day2

+++ database/db_user.py
@@ -2,6 +2,7 @@ from sqlalchemy.orm.session import Session
 from schemas import UserBase
 from database.models import DbUser
 from database.hash import Hash
+from fastapi import HTTPException, status
 
 
 def create_users(db: Session, request: UserBase):
@@ -24,11 +25,18 @@ def get_all_users(db: Session):
 
 
 def get_one_user(db: Session, id: int):
-    return db.query(DbUser).filter(DbUser.id == id).first()
+    user = db.query(DbUser).filter(DbUser.id == id).first()
+    if not user:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
+    return user
 
 
 def update_user(id: int, db: Session, request: UserBase):
     user = db.query(DbUser).filter(DbUser.id == id)
+    if not user.first():
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
     new_data = DbUser(
         username=request.username,
         email=request.email,
@@ -44,7 +52,10 @@ def update_user(id: int, db: Session, request: UserBase):
 
 
 def delete_user(id: int, db: Session):
-    user = db.query(DbUser).filter(DbUser.id == id).first()
-    db.delete(user)
-    db.commit()
-    return {"deleted": id}
+    if user := db.query(DbUser).filter(DbUser.id == id).first():
+        db.delete(user)
+        db.commit()
+        return {"deleted": id}
+    else:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+                            detail=f"User with id {id} not found")
